green_truth_auditor/requirements.txt
flask
flask-cors
torch
torchvision
transformers
sentence-transformers
beautifulsoup4
requests
pillow
easyocr
pandas
numpy
accelerate